TST: isolating unit tests from test_tools_matrix
